Consolidate the Moq1001 rule to use the MockExpressionHelper to determine the MockBehavior.
