+++ src/Moq.Analyzers/Analyzers/MockInstanceShouldBeStrictBehaviorAnalyzer.cs
@@ -53,73 +53,10 @@ namespace PosInformatique.Moq.Analyzers
                 return;
             }
 
-            // Check that the "new Mock<I>()" statement have at least one argument (else Strict is missing...).
-            if (objectCreation.ArgumentList is null)
+            if (!MockExpressionHelper.IsStrictBehavior(moqSymbols, context.SemanticModel, objectCreation))
             {
                 var diagnostic = Diagnostic.Create(Rule, objectCreation.GetLocation());
                 context.ReportDiagnostic(diagnostic);
-
-                return;
-            }
-
-            var firstArgument = objectCreation.ArgumentList.Arguments.FirstOrDefault();
-
-            if (firstArgument is null)
-            {
-                var diagnostic = Diagnostic.Create(Rule, objectCreation.GetLocation());
-                context.ReportDiagnostic(diagnostic);
-
-                return;
-            }
-
-            // Gets the first argument of "new Mock<I>(...)" and ensures it is a MemberAccessExpressionSyntax
-            // (because we searching for MockBehavior.Strict).
-            if (firstArgument.Expression is not MemberAccessExpressionSyntax memberAccessExpression)
-            {
-                var diagnostic = Diagnostic.Create(Rule, objectCreation.GetLocation());
-                context.ReportDiagnostic(diagnostic);
-
-                return;
-            }
-
-            // Find the "MockBehavior" type in the semantic model from Moq library.
-            var mockBehaviorType = context.Compilation.GetTypeByMetadataName("Moq.MockBehavior");
-
-            if (mockBehaviorType is null)
-            {
-                // Moq not installed (dependency of the Moq package missing), so we stop analysis.
-                return;
-            }
-
-            // Check that the "memberAccessExpression.Expression" is applied on the Moq MockBehavior type.
-            var firstArgumentType = context.SemanticModel.GetSymbolInfo(memberAccessExpression.Expression);
-
-            if (!SymbolEqualityComparer.Default.Equals(firstArgumentType.Symbol, mockBehaviorType))
-            {
-                var diagnostic = Diagnostic.Create(Rule, firstArgument.GetLocation());
-                context.ReportDiagnostic(diagnostic);
-
-                return;
-            }
-
-            // Find the Strict field in the semantic model
-            var strictField = mockBehaviorType.GetMembers("Strict").SingleOrDefault();
-
-            if (strictField is null)
-            {
-                // The field Strict seam to not exists (dependency of the Moq package missing ? Or something wrong ?), so we stop analysis.
-                return;
-            }
-
-            // Check that the memberAccessExpression.Name reference the Strict field
-            var firstArgumentField = context.SemanticModel.GetSymbolInfo(memberAccessExpression.Name);
-
-            if (!SymbolEqualityComparer.Default.Equals(firstArgumentField.Symbol, strictField))
-            {
-                var diagnostic = Diagnostic.Create(Rule, memberAccessExpression.Name.GetLocation());
-                context.ReportDiagnostic(diagnostic);
-
-                return;
             }
         }
     }
