+++ tests/Moq.Analyzers.Tests/Analyzers/MockInstanceShouldBeStrictBehaviorAnalyzerTest.cs
@@ -82,7 +82,7 @@ namespace PosInformatique.Moq.Analyzers.Tests
                     {
                         public void TestMethod()
                         {
-                            var mock1 = new Mock<I>(MockBehavior.[|Loose|]);
+                            var mock1 = [|new Mock<I>(MockBehavior.Loose)|];
                         }
                     }
 
@@ -189,7 +189,7 @@ namespace PosInformatique.Moq.Analyzers.Tests
         }
 
         [Fact]
-        public async Task NoMoqNamespace()
+        public async Task NoMoqLibrary()
         {
             var source = @"
                 namespace ConsoleApplication1
@@ -214,48 +214,6 @@ namespace PosInformatique.Moq.Analyzers.Tests
                     public class Mock<T>
                     {
                         public Mock(MockBehavior _) { }
-                    }
-
-                    public enum MockBehavior { Strict, Loose }
-                }";
-
-            await Verify.VerifyAnalyzerAsync(source);
-        }
-
-        [Fact]
-        public async Task MockBehaviorTypeNotFromMoq()
-        {
-            var source = @"
-                namespace ConsoleApplication1
-                {
-                    using Moq;
-                    using OtherNamespace;
-
-                    public class TestClass
-                    {
-                        public void TestMethod()
-                        {
-                            var mock1 = new Mock<I>([|NotMockBehavior.Strict|]);
-                        }
-                    }
-
-                    public interface I
-                    {
-                    }
-                }
-
-                namespace OtherNamespace
-                {
-                    public enum NotMockBehavior { Strict, Loose }
-                }
-
-                namespace Moq
-                {
-                    using OtherNamespace;
-
-                    public class Mock<T>
-                    {
-                        public Mock(NotMockBehavior _) { }
                     }
 
                     public enum MockBehavior { Strict, Loose }
